+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Dashboard: Search for the error on multiple search engines
+- API: Add response actions
+
 ## [0.2.1] - 2019-11-24
 
 ### Added
 
-- Extend / pretty-print JavaScript-Errors with [`stacktracey`]
-- Ability to show last 5 values per Watcher
+- Dashboard: Extend / pretty-print JavaScript-Errors with [`stacktracey`]
+- Dashboard: Ability to show last 5 values per Watcher
 
 [`stacktracey`]: https://www.npmjs.com/package/stacktracey